lstm

+++ 1.olistkaggle - Código 2 classes ptbr/algoritmos/LSTM_embeddings.py
@@ -12,8 +12,8 @@ from tqdm import tqdm
 with open('../corpus_embeddingsLSTM.pkl', 'rb') as f:
     data = pickle.load(f)
 
-X = torch.tensor(np.array(data['embedding']))# Embeddings
-y = torch.tensor(np.array(data['polarity']))   # Classes
+X = torch.tensor(np.array(data['embedding']), dtype=torch.float32)# Embeddings
+y = torch.tensor(np.array(data['polarity']), dtype=torch.float32)   # Classes
 
 # 2. Separar 15% dos dados para TESTE
 X_temp, X_test, y_temp, y_test = train_test_split(X, y, test_size=0.15, stratify=y, random_state=42)
